requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
distro==1.9.0
entrypoints==0.4
Faker==22.1.0
favicon==0.7.0
fonttools==4.47.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.0
langchain-community==0.0.11
langchain-core==0.1.9
langsmith==0.0.79
lxml==5.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.7.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.25.1
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
st-annotated-text==4.0.1
st-pages==0.4.5
streamlit==1.29.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
yarl==1.9.4
zipp==3.17.0
